test(arnes): split measurement ingestion coverage
